+++ test/helpers/jobs_helper_test.rb
@@ -1,6 +1,7 @@
 require 'test_helper'
 
 class JobsHelperTest < ActionView::TestCase
+  include ApplicationHelper
   include JobsHelper
 
   setup do